fix windows builds

+++ src/vera/vera_video.cpp
@@ -1204,7 +1204,7 @@ void fx_affine_prefetch(void)
 
 	if (affine_x_tile >= fx_affine_map_size || affine_y_tile >= fx_affine_map_size) {
 		// We clipped, return value for tile 0
-		address = fx_affine_tile_base + (affine_y_sub_tile << (3 - fx_4bit_mode)) + (affine_x_sub_tile >> fx_4bit_mode);
+		address = fx_affine_tile_base + (affine_y_sub_tile << (3 - fx_4bit_mode)) + (affine_x_sub_tile >> (uint8_t)fx_4bit_mode);
 		if (fx_4bit_mode) fx_nibble_bit[1] = 0;
 	} else {
 		// Get the address within the tile map
@@ -1213,7 +1213,7 @@ void fx_affine_prefetch(void)
 		uint8_t affine_tile_idx = vera_video_space_read(address);
 		address = fx_affine_tile_base + (affine_tile_idx << (6 - fx_4bit_mode));
 		// Now add the sub-tile address
-		address += (affine_y_sub_tile << (3 - fx_4bit_mode)) + (affine_x_sub_tile >> fx_4bit_mode);
+		address += (affine_y_sub_tile << (3 - fx_4bit_mode)) + (affine_x_sub_tile >> (uint8_t)fx_4bit_mode);
 		if (fx_4bit_mode) fx_nibble_bit[1] = affine_x_sub_tile & 1;
 	}
 	io_addr[1] = address;
